SOAP action value is not needed in order to invoke the service.
There are some web services declaring operations in their WSDLs with
soapAction attribute empty. This seems to still be WS-I compliant so
we are going to go ahead and generate operations anyway. Also, the
SOAP envelope carries all the information required by the server
in order to know exactly what operation is being invoked. And, servers
should not rely in the value of the SOAPAction HTTP header as per point
3.4.3 in http://www.ws-i.org/profiles/basicprofile-1.1-2004-08-24.html