Allow clients to add SOAPHeader in SOAPClient.
In order to allow the SOAPClient to send a
SOAPHeader without modification of generated code,
a new method SetHeader was added to both the
SOAPClient and to the generated operation.
The header, once set, is preserved between
invocations of other methods in the same instance.

This allows client code to call SetHeader()
with the desired header, including nil to remove
it from future calls in the same instance.

The SOAPHeader usually holds authentication info,
and we can benefit from reusing already set
credentials between invocations.