Added basic support for WSS Security header: new structs WSSSecurityHeader, WSSUsernameToken and its child structs WSSUsername and WSSPassword. XML namespace problems are dealt with by using vania-pooh's workaround:  https://github.com/golang/go/issues/9519#issuecomment-252196382
Convenience function "NewWSSSecurityHeader" to ease header creation (the header is not automatically added to the SOAP client).

"SetHeader" method is kept for compatibility. There is a new "AddHeader" method to add headers to the SOAP call. There is now support for several headers instead of only one.

Cosmetic changes in header_tmpl.go to sort import names alphabetically.

Sample code:

   ...
   hdr := myservice.NewWSSSecurityHeader("sampleuser", "samplepassword", "1")  // "1" is for 'mustUnderstand' attribute; empty if not specified
   myserviceClient.AddHeader(hdr)
   ...