Fix dereference of schemaLocation attribute on <include>
When schema location values (i.e. schemaLocation attributes on <include>, <redefine>, <override>, and <import> in schema documents, or xsi:schemaLocation and xsi:noNamespaceSchemaLocation attributes in instance documents) are dereferenced and the values are relative references, then the [base URI] of the [owner element] must be used to resolve the relative references.