Only add an XMLName field to structs if it is used by a top-level element
This was originally implemented with PR #109 (issue #73) in commit
dd8f1cae4f65c55c7ddca89a21925dd9d85ad14e.

I am not entirely sure what the original intention was. The logic
in the template (in conjunction with the findNameByType template
function) attempts to only add the XMLName field to structs if it
finds a top-level element of that struct type. If this is the case,
it takes the name of the found element for the tag of the XMLName
field.

Because it compared the output of the findNameByType function to the
generated struct type name (which has gone through replaceReservedWords
and makePublic), it added the XMLName field also to structs that do
not have a top-level element of that struct type. Simple examples are
struct type names starting with a lower case letter which are then
made public with makePublic. In this case, the upper case struct type
name does not match the lower case type name given to findNameByType,
which is returned unmodified, if no top-level element is found.

The reason why it is important not to add XMLName field to all struct
type that are not used at the top-level, is that the tags can easily
generate name mismatchs due to what is described like this in the Go
doc https://pkg.go.dev/encoding/xml:
"If the XML name for a struct field is defined by both the field tag
and the struct's XMLName field, the names must match."

Errors could look like these:
- "xml: name \"order\" in tag of example.AckType.Order
   conflicts with name \"orderType\" in *example.OrderType.XMLName"
- "expected element type <historyType> but have <history>"