Avoid XML name conflicts in XML tags between field tag and struct's XMLName field
This fixes #224.
It was first addressed with PR #109 (issue #73) in commit
dd8f1cae4f65c55c7ddca89a21925dd9d85ad14e. However this does not
address all possible name mismatch scenarios.

Name mismatches/conflicts may occur if a complexType struct is
used as type by an XML element with a different name.
The Go doc https://pkg.go.dev/encoding/xml says:
"If the XML name for a struct field is defined by both the field
 tag and the struct's XMLName field, the names must match."

To solve this problem, the logic so far used the name of the
top-level element for the struct's XMLName field tag. This is fine
as long as only a single element is of that complexType (or
multiple elements with identical names for that matter).

However this logic breaks down if multiple elements with different
names are of that complexType. In this case the logic would
arbitrarly take the first element's name for the struct's XMLName
field tag. All other elements with different names would then
conflicts with that tag.

Errors could look like these:
- "xml: name \"order\" in tag of example.AckType.Order
   conflicts with name \"orderType\" in *example.OrderType.XMLName"
- "expected element type <historyType> but have <history>"

To solve this problem an additional "traverse" mode is introduced
to find element names by type name. Three cases are distinguished:
1. If a single element or multiple elements with identical name
   are found, that name is used for the struct's XMLName field tag
2. If multiple elements with different names are found, the
   XMLName field is omitted from the struct
3. If no elements are found, the  XMLName field is omitted from
   the struct

For this to work, two additional changes are required:
- for complexTypes also traverse elements in complexContent:
  - ComplexContent.Extension.Sequence
  - ComplexContent.Extension.Choice
  - ComplexContent.Extension.SequenceChoice
- Sequence in XSDExtension must be of pointer type []*XSDElement